requirements.txt
pyotp==2.6.0
python-dateutil==2.8.2
python-dotenv==0.19.0
requests==2.26.0
selenium==3.141.0
telegram-send==0.25
urllib3==1.26.6